Move NotificationProperties
