+++ publisher-web/app/views/datasets/show.scala.html
@@ -4,7 +4,7 @@
 
 @import nl.idgis.publisher.domain.web._
 @import nl.idgis.publisher.domain.job.ConfirmNotificationResult
-@import nl.idgis.publisher.domain.job.load.NotificationProperties
+@import nl.idgis.publisher.domain.job.NotificationProperties
 @import nl.idgis.publisher.domain.service.ColumnDiffOperation
 
 @categoryRefLink(cat: Category, cls: String = null) = {
